+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Static methods are now supported on classes, elements, and mixins.
 
 ## [0.3.1] - 2017-12-15
 - Convert Closure `Object` to TypeScript `object`.